pyproject.toml
[project]
name = "psl-ai-engineer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate>=1.13.0",
    "google-generativeai>=0.8.0",
    "chromadb>=1.5.9",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.3.0",
    "langchain-community>=0.4.1",
    "numpy>=2.4.4",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pydantic>=2.13.4",
    "pymupdf>=1.27.2.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "reportlab>=4.5.1",
    "rich>=15.0.0",
    "sentence-transformers>=5.5.0",
    "transformers>=5.8.1",
    "typer>=0.25.1",
    "uvicorn>=0.46.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "easyocr>=1.7.2",
    "python-multipart>=0.0.28",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
name = "pytorch-cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
